move edit actions to separate files; fix color in code editor
