+++ packages/vtv/src/model/actions/index.js
@@ -1,1 +1,3 @@
 export { default as rename } from './rename'
+export { default as edit } from './edit'
+export { default as editJson } from './editJson'
