+++ packages/vtv/src/model/util/index.js
@@ -6,3 +6,4 @@ export { default as clearStateProperty } from './clearStateProperty'
 export { default as getState } from './getState'
 export { default as getChildState } from './getChildState'
 export { default as getNestedState } from './getNestedState'
+export { default as getDraftUpdate } from './getDraftUpdate'
